# airflow/api_connexion/openapi/v1.yaml
...
openapi: 3.0.3
info:
  title: "Airflow API (Stable)"
  description: |
    # Overview

    To facilitate management, Apache Airflow supports a range of REST API endpoints across its
    objects.
    This section provides an overview of the API design, methods, and supported use cases.

    Most of the endpoints accept `JSON` as input and return `JSON` responses.
    This means that you must usually add the following headers to your request:
    ```
    Content-type: application/json
    Accept: application/json
    ```

    ## Resources

    The term `resource` refers to a single type of object in the Airflow metadata. An API is broken up by its
    endpoint's corresponding resource.
    The name of a resource is typically plural and expressed in camelCase. Example: `dagRuns`.

    Resource names are used as part of endpoint URLs, as well as in API parameters and responses.

    ## CRUD Operations

    The platform supports **C**reate, **R**ead, **U**pdate, and **D**elete operations on most resources.
    You can review the standards for these operations and their standard parameters below.

    Some endpoints have special behavior as exceptions.

    ### Create

    To create a resource, you typically submit an HTTP `POST` request with the resource's required metadata
    in the request body.
    The response returns a `201 Created` response code upon success with the resource's metadata, including
    its internal `id`, in the response body.

    ### Read

...

...

...

...
    ```
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /example_invalidate_dagbag_cache:
    post:
      # TODO: this is an example do not MERGE THIS TO MAIN
      summary: example invalidation of a dagbag cache
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: example_invalidate_dagbag_cache
      tags: [DAG]
      responses:
        '200':
          description: Success.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
  /dags:
    get:
      summary: List DAGs
      description: >
       List DAGs in the database.

       `dag_id_pattern` can be set to match dags of a specific pattern
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: get_dags
      tags: [DAG]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/OrderBy'
        - $ref: '#/components/parameters/FilterTags'
        - $ref: '#/components/parameters/OnlyActive'
        - $ref: '#/components/parameters/Paused'
        - name: dag_id_pattern
          in: query
          schema:
            type: string
          required: false
          description: |
            If set, only return DAGs with dag_ids matching this pattern.
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAGCollection'
        '401':
          $ref: '#/components/responses/Unauthenticated'

    patch:
      summary: Update DAGs
      description: >
       Update DAGs of a given dag_id_pattern using UpdateMask.

       This endpoint allows specifying `~` as the dag_id_pattern to update all DAGs.

       *New in version 2.3.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: patch_dags
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/FilterTags'
        - $ref: '#/components/parameters/UpdateMask'
        - $ref: '#/components/parameters/OnlyActive'
        - name: dag_id_pattern
          in: query
          schema:
            type: string
          required: true
          description: |
            If set, only update DAGs with dag_ids matching this pattern.
      tags: [DAG]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DAG'
            example:
              is_paused: true
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAGCollection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
  /dags/{dag_id}:
    parameters:
      - $ref: '#/components/parameters/DAGID'

    get:
      summary: Get basic information about a DAG
      description: >
        Presents only information available in database (DAGModel).

        If you need detailed information, consider using GET /dags/{dag_id}/details.
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: get_dag
      tags: [DAG]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAG'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update a DAG
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: patch_dag
      parameters:
        - $ref: '#/components/parameters/UpdateMask'
      tags: [DAG]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DAG'
            example:
              is_paused: true
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAG'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a DAG
      description: |
        Deletes all metadata related to the DAG, including finished DAG Runs and Tasks.
        Logs are not deleted. This action cannot be undone.

        *New in version 2.2.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: delete_dag
      tags: [DAG]
      responses:
        '204':
          description: Success.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AlreadyExists'
...
  /dags/{dag_id}/details:
    parameters:
      - $ref: '#/components/parameters/DAGID'

    get:
      summary: Get a simplified representation of DAG
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: get_dag_details
      description: >
        The response contains many DAG attributes, so the response can be large.
        If possible, consider using GET /dags/{dag_id}.
      tags: [DAG]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DAGDetail'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  schemas:
...
    DAG:
      type: object
      description: DAG
      properties:
        dag_id:
          type: string
          readOnly: true
          description: The ID of the DAG.
        root_dag_id:
          type: string
          readOnly: true
          nullable: true
          description: If the DAG is SubDAG then it is the top level DAG identifier. Otherwise, null.
        is_paused:
          type: boolean
          nullable: true
          description: Whether the DAG is paused.
        is_active:
          description: |
            Whether the DAG is currently seen by the scheduler(s).

            *New in version 2.1.1*

            *Changed in version 2.2.0*&#58; Field is read-only.
          nullable: true
          readOnly: true
          type: boolean
        is_subdag:
          description: Whether the DAG is SubDAG.
          type: boolean
          readOnly: true
        last_parsed_time:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            The last time the DAG was parsed.

            *New in version 2.3.0*
        last_pickled:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            The last time the DAG was pickled.

            *New in version 2.3.0*
        last_expired:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            Time when the DAG last received a refresh signal
            (e.g. the DAG's "refresh" button was clicked in the web UI)

            *New in version 2.3.0*
        scheduler_lock:
          type: boolean
          readOnly: true
          nullable: true
          description: |
            Whether (one of) the scheduler is scheduling this DAG at the moment

            *New in version 2.3.0*
        pickle_id:
          type: string
          readOnly: true
          nullable: true
          description: |
            Foreign key to the latest pickle_id

            *New in version 2.3.0*
        default_view:
          type: string
          nullable: true
          readOnly: true
          description: |
            Default view of the DAG inside the webserver

            *New in version 2.3.0*
        fileloc:
          description: The absolute path to the file.
          type: string
          readOnly: true
        file_token:
          type: string
          readOnly: true
          description: >
            The key containing the encrypted path to the file. Encryption and decryption take place only on
            the server. This prevents the client from reading an non-DAG file. This also ensures API
            extensibility, because the format of encrypted data may change.
        owners:
          type: array
          items:
            type: string
          readOnly: true
        description:
          type: string
          readOnly: true
          nullable: true
          description: >
            User-provided DAG description, which can consist of several sentences or paragraphs that
            describe DAG contents.
        schedule_interval:
          $ref: '#/components/schemas/ScheduleInterval'
        timetable_description:
          type: string
          readOnly: true
          nullable: true
          description: |
            Timetable/Schedule Interval description.

            *New in version 2.3.0*
        tags:
          description: List of tags.
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Tag'
          readOnly: true
        max_active_tasks:
          type: integer
          nullable: true
          readOnly: true
          description: |
            Maximum number of active tasks that can be run on the DAG

            *New in version 2.3.0*
        max_active_runs:
          type: integer
          nullable: true
          readOnly: true
          description: |
            Maximum number of active DAG runs for the DAG

            *New in version 2.3.0*
        has_task_concurrency_limits:
          type: boolean
          nullable: true
          readOnly: true
          description: |
            Whether the DAG has task concurrency limits

            *New in version 2.3.0*
        has_import_errors:
          type: boolean
          nullable: true
          readOnly: true
          description: |
            Whether the DAG has import errors

            *New in version 2.3.0*
        next_dagrun:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            The logical date of the next dag run.

            *New in version 2.3.0*
        next_dagrun_data_interval_start:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            The start of the interval of the next dag run.

            *New in version 2.3.0*
        next_dagrun_data_interval_end:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            The end of the interval of the next dag run.

            *New in version 2.3.0*
        next_dagrun_create_after:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: |
            Earliest time at which this ``next_dagrun`` can be created.

            *New in version 2.3.0*
    DAGCollection:
      description: |
        Collection of DAGs.

        *Changed in version 2.1.0*&#58; 'total_entries' field is added.
      type: object
      allOf:
        - type: object
          properties:
            dags:
              type: array
              items:
                $ref: '#/components/schemas/DAG'
        - $ref: '#/components/schemas/CollectionInfo'
...
    DAGDetail:
      description: |
        DAG details.

        For details see:
        [airflow.models.DAG](https://airflow.apache.org/docs/apache-airflow/stable/_api/airflow/models/index.html#airflow.models.DAG)
      allOf:
        - $ref: '#/components/schemas/DAG'
        - type: object
          properties:
            timezone:
              $ref: '#/components/schemas/Timezone'
            catchup:
              type: boolean
              readOnly: true
            orientation:
              type: string
              readOnly: true
            concurrency:
              type: number
              readOnly: true
            start_date:
              type: string
              format: 'date-time'
              readOnly: true
              nullable: true
              description: |
                The DAG's start date.

                *Changed in version 2.0.1*&#58; Field becomes nullable.
            dag_run_timeout:
              nullable: true
              $ref: '#/components/schemas/TimeDelta'
            doc_md:
              type: string
              readOnly: true
              nullable: true
            default_view:
              type: string
              readOnly: true
            params:
              type: object
              readOnly: true
              description: |
                User-specified DAG params.

                *New in version 2.0.1*
            end_date:
              type: string
              format: 'date-time'
              readOnly: true
              nullable: true
              description: |
                The DAG's end date.

                *New in version 2.3.0*.
            is_paused_upon_creation:
              type: boolean
              readOnly: true
              nullable: true
              description: |
                Whether the DAG is paused upon creation.

                *New in version 2.3.0*
            last_parsed:
              type: string
              format: date-time
              nullable: true
              readOnly: true
              description: |
                The last time the DAG was parsed.

                *New in version 2.3.0*
            template_search_path:
              type: array
              nullable: true
              items:
                type: string
              description: |
                The template search path.

                *New in version 2.3.0*
            render_template_as_native_obj:
              type: boolean
              nullable: true
              readOnly: true
              description: |
                Whether to render templates as native Python objects.

                *New in version 2.3.0*
...
    ScheduleInterval:
      description: |
        Schedule interval. Defines how often DAG runs, this object gets added to your latest task instance's
        execution_date to figure out the next schedule.
      nullable: true
      readOnly: true
      anyOf:
        - $ref: '#/components/schemas/TimeDelta'
        - $ref: '#/components/schemas/RelativeDelta'
        - $ref: '#/components/schemas/CronExpression'
      discriminator:
        propertyName: __type
    TimeDelta:
      description: Time delta
      type: object
      required:
        - __type
        - days
        - seconds
        - microseconds
      properties:
        __type: {type: string}
        days: {type: integer}
        seconds: {type: integer}
        microseconds: {type: integer}
    RelativeDelta:
      description: Relative delta
      # TODO: Why we need these fields?
      type: object
      required:
        - __type
        - years
        - months
        - days
        - leapdays
        - hours
        - minutes
        - seconds
        - microseconds
        - year
        - month
        - day
        - hour
        - minute
        - second
        - microsecond
      properties:
        __type: {type: string}
        years: {type: integer}
        months: {type: integer}
        days: {type: integer}
        leapdays: {type: integer}
        hours: {type: integer}
        minutes: {type: integer}
        seconds: {type: integer}
        microseconds: {type: integer}
        year: {type: integer}
        month: {type: integer}
        day: {type: integer}
        hour: {type: integer}
        minute: {type: integer}
        second: {type: integer}
        microsecond: {type: integer}
    CronExpression:
      description: Cron expression
      type: object
      required:
        - __type
        - value
      properties:
        __type: {type: string}
        value: {type: string}
      nullable: true
    Timezone:
      type: string
    Tag:
      description: Tag
      # Object to maintain extensibility
      type: object
      properties:
        name:
          type: string
...
    Error:
      description: |
        [RFC7807](https://tools.ietf.org/html/rfc7807) compliant response.
      type: object
      properties:
        type:
          type: string
          description: |
            A URI reference [RFC3986] that identifies the problem type. This specification
            encourages that, when dereferenced, it provide human-readable documentation for
            the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: number
          description: The HTTP status code generated by the API server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: |
            A URI reference that identifies the specific occurrence of the problem. It may or may
            not yield further information if dereferenced.
      required:
        - type
        - title
        - status
    CollectionInfo:
      description: Metadata about collection.
      type: object
      properties:
        total_entries:
          type: integer
          description: |
            Count of total objects in the current result set before pagination parameters
            (limit, offset) are applied.

    # Enums
...
  parameters:
...
    PageOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    PageLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 100
      description: The numbers of items to return.

    # Database entity fields
...
    DAGID:
      in: path
      name: dag_id
      schema:
        type: string
      required: true
      description: The DAG ID.
...
    FilterTags:
      in: query
      name: tags
      schema:
        type: array
        items:
          type: string
      description: |
          List of tags to filter results.

          *New in version 2.2.0*
...
    OrderBy:
      in: query
      name: order_by
      schema:
        type: string
      required: false
      description: |
        The name of the field to order the results by.
        Prefix a field name with `-` to reverse the sort order.

        *New in version 2.1.0*
    OnlyActive:
      in: query
      name: only_active
      schema:
        type: boolean
        default: true
      required: false
      description: |
        Only filter active DAGs.

        *New in version 2.1.1*
...
    Paused:
      in: query
      name: paused
      schema:
        type: boolean
      required: false
      description: |
        Only filter paused/unpaused DAGs. If absent or null, it returns paused and unpaused DAGs.

        *New in version 2.6.0*

    # Other parameters
...
    UpdateMask:
      in: query
      name: update_mask
      schema:
        type: array
        items:
          type: string
      description: |
        The fields to update on the resource. If absent or empty, all modifiable fields are updated.
        A comma-separated list of fully qualified names of fields.
      style: form
      explode: false

  # Reusable request bodies
...
  responses:
...
    'BadRequest':
      description: Client specified an invalid argument.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 401
    'Unauthenticated':
      description: Request not authenticated due to missing, invalid, authentication info.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 403
    'PermissionDenied':
      description: Client does not have sufficient permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 404
    'NotFound':
      description: A specified resource is not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 405
...
    'AlreadyExists':
      description: An existing resource conflicts with the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    # 500
...
  securitySchemes:
    Basic:
      type: http
      scheme: basic
    GoogleOpenId:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
    Kerberos:
      type: http
      scheme: negotiate

# The API will provide support for plugins to support various authorization mechanisms.
# Detailed information will be available in the plugin specification.
security: []
...